create tmp 003_*.lua test file; will migrate manually
The migration script produced a faulty test file; I will migrate the
test suite by hand to track down the problem.
